requirements.txt
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.2
Flask==3.0.3
greenlet==3.1.1
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
pyodbc==5.1.0
SQLAlchemy==2.0.35
uvicorn==0.31.1
python-multipart==0.0.12
pydantic==2.9.2
starlette==0.40.0
